ZO-1408: Move 3rd party tests to separate module.
